.claude/skills/deep-paper-search/requirements.txt
llama-index-core>=0.12.0
llama-index-readers-papers>=0.3.0
llama-index-embeddings-huggingface>=0.5.0
